total deposit/borrow value charts, new chart styles
